+++ src/apsg/plotting/__init__.py
@@ -1,14 +1,15 @@
 # -*- coding: utf-8 -*-
 
-from apsg.plotting._stereonet import StereoNet, quicknet
-from apsg.plotting._stereogrid import StereoGrid
-from apsg.plotting._roseplot import RosePlot
-from apsg.plotting._fabricplot import VollmerPlot, RamsayPlot, FlinnPlot, HsuPlot
+from apsg.plotting._fabricplot import FlinnPlot, HsuPlot, RamsayPlot, VollmerPlot
 from apsg.plotting._plot_artists import (
-    StereoNetArtistFactory,
-    RosePlotArtistFactory,
     FabricPlotArtistFactory,
+    RosePlotArtistFactory,
+    StereoNetArtistFactory,
 )
+from apsg.plotting._roseplot import RosePlot
+from apsg.plotting._stereogrid import StereoGrid
+from apsg.plotting._stereonet import StereoNet, quicknet
+from apsg.plotting._styles import StereoNetStyleFactory
 
 __all__ = (
     "StereoNet",
@@ -22,4 +23,5 @@ __all__ = (
     "StereoNetArtistFactory",
     "RosePlotArtistFactory",
     "FabricPlotArtistFactory",
+    "StereoNetStyleFactory",
 )