+++ src/apsg/plotting/_fabricplot.py
@@ -7,6 +7,7 @@ from matplotlib.patches import Polygon
 from apsg.config import apsg_conf
 from apsg.feature import feature_from_json
 from apsg.plotting._plot_artists import FabricPlotArtistFactory
+from apsg.plotting._styles import FabricPlotStyle
 
 __all__ = ["VollmerPlot", "RamsayPlot", "FlinnPlot", "HsuPlot"]
 
@@ -125,6 +126,30 @@ class FabricPlot(object):
         self._render()
         self.fig.savefig(filename, **kwargs)
         plt.close(0)
+
+    ########################################
+    # STYLED PLOTTING                      #
+    ########################################
+
+    def plot(self, style, *args):
+        """
+        Plot features using apsg styles
+
+        Args:
+            style: apsg plotting style. See fabricplot_styles
+            *arg: any number of features to be plotted
+
+        Note:
+            Features in args are automatically filtered by style to accept only compatible features
+
+        """
+        assert issubclass(
+            type(style), FabricPlotStyle
+        ), "Style must FabricPlotStyle object"
+
+        artist = style.create_artist(*args)
+        if len(artist.args) > 0:
+            self._artists.append(artist)
 
 
 class VollmerPlot(FabricPlot):