+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-### 1.4.1 (master)
+### 1.5.0 (master)
+ * StereoNet point method added as replacement for both line and pole
+ * StereoNet plot method using styles implemented
  * Rotation classes derived from deformation gradient class implemented
  * align method added to featureset to find best rotation
  * Watson statistics added